refactored integration tests + add slo workloads paths
