+++ tests/integration/README.md
@@ -1,0 +1,8 @@
+# Package `integration`
+
+Package `integration` contains only integration tests for `ydb-go-sdk`. All test files must have build tag
+```go
+//go:build !fast
+// +build !fast
+```
+for run this test files as integration tests int github action `integration`.
