+++ tests/slo/database/sql/README.md
@@ -1,0 +1,1 @@
+Workload application
