+++ tests/slo/native/README.md
@@ -1,0 +1,1 @@
+Workload application
